C++: Fix compiler warnings from gcc-4.7 on Fedora 18
Clang didn't complain about these. Note that some of the error messages
for not-clang in the tests have changed, likely due to C++. The old ones
are commented out for now; this part needs to be tested with an older
gcc to see if it's consistent.

+++ src/libqof/qof/qofmath128.c
@@ -197,7 +197,7 @@ div128 (qofint128 n, gint64 d)
         remainder <<= 1;
         if (sbit) remainder |= 1;
         quotient = shiftleft128 (quotient);
-        if (remainder >= d)
+        if (remainder >= static_cast<unsigned int64_t>(d))
         {
             remainder -= d;
             quotient.lo |= 1;
